Moves production to the new host and closes the findings of a four-round
audit: code review across seven areas, fuzzing of every parser reachable from
the network, a sweep for SQL and output injection, dependency scanning, and
measured behaviour under load and dependency failure.

The server-side fixes are already live; this brings the branch into master.

+++ internal/server/database/sqlc/querier.go
@@ -87,7 +87,7 @@ type Querier interface {
 	GetHistoryEntryByID(ctx context.Context, arg GetHistoryEntryByIDParams) (UserHistory, error)
 	GetHistoryStats(ctx context.Context, userID int64) (GetHistoryStatsRow, error)
 	GetLatestAuditLogByUserAndAction(ctx context.Context, arg GetLatestAuditLogByUserAndActionParams) (AuditLog, error)
-	GetNextInvoiceID(ctx context.Context) (int32, error)
+	GetNextInvoiceID(ctx context.Context) (int64, error)
 	GetPaymentByID(ctx context.Context, id int64) (Payment, error)
 	GetPaymentByInvoiceID(ctx context.Context, invoiceID int64) (Payment, error)
 	GetPendingPaymentsBySubscriptionID(ctx context.Context, subscriptionID pgtype.Int8) ([]Payment, error)
